Refactor error handling and request headers
